+++ include/alpaka/alpaka.hpp
@@ -139,14 +139,15 @@
 #include "alpaka/mem/alloc/AllocCpuAligned.hpp"
 #include "alpaka/mem/alloc/AllocCpuNew.hpp"
 #include "alpaka/mem/alloc/Traits.hpp"
-#include "alpaka/mem/buf/BufCpu.hpp"
-#include "alpaka/mem/buf/BufCpuSycl.hpp"
-#include "alpaka/mem/buf/BufCudaRt.hpp"
-#include "alpaka/mem/buf/BufFpgaSyclIntel.hpp"
-#include "alpaka/mem/buf/BufGenericSycl.hpp"
-#include "alpaka/mem/buf/BufGpuSyclIntel.hpp"
-#include "alpaka/mem/buf/BufHipRt.hpp"
 #include "alpaka/mem/buf/Traits.hpp"
+#include "alpaka/mem/buf/cpu/CpuBuf.hpp"
+#include "alpaka/mem/buf/sycl/BufCpuSycl.hpp"
+#include "alpaka/mem/buf/sycl/BufFpgaSyclIntel.hpp"
+#include "alpaka/mem/buf/sycl/BufGpuSyclIntel.hpp"
+#include "alpaka/mem/buf/sycl/ConstBufGenericSycl.hpp"
+#include "alpaka/mem/buf/uniformCudaHip/BufCudaRt.hpp"
+#include "alpaka/mem/buf/uniformCudaHip/BufHipRt.hpp"
+#include "alpaka/mem/buf/uniformCudaHip/BufUniformCudaHipRt.hpp"
 #include "alpaka/mem/fence/MemFenceCpu.hpp"
 #include "alpaka/mem/fence/MemFenceCpuSerial.hpp"
 #include "alpaka/mem/fence/MemFenceGenericSycl.hpp"
