+++ include/alpaka/mem/buf/cpu/MutBufCpu.hpp
@@ -1,0 +1,40 @@
+/* Copyright 2025 Anton Reinhard
+ * SPDX-License-Identifier: MPL-2.0
+ */
+
+#pragma once
+
+#include "alpaka/dev/Traits.hpp"
+#include "alpaka/mem/buf/Traits.hpp"
+#include "alpaka/mem/view/ViewAccessOps.hpp"
+#include "alpaka/vec/Vec.hpp"
+
+#include <functional>
+#include <memory>
+#include <type_traits>
+#include <utility>
+
+namespace alpaka
+{
+    //! The CPU memory buffer template implementing muting accessors.
+    template<
+        template<typename, typename, typename> class TBuf,
+        typename TBufImpl,
+        typename TDev,
+        typename TElem,
+        typename TDim,
+        typename TIdx>
+    class MutCpuBuf : public internal::ViewAccessOps<MutCpuBuf<TBuf, TBufImpl, TDev, TElem, TDim, TIdx>>
+    {
+    public:
+        template<typename TExtent, typename Deleter>
+        ALPAKA_FN_HOST MutCpuBuf(TDev const& dev, TElem* const pMem, Deleter deleter, TExtent const& extent)
+            : m_spBufImpl{std::make_shared<TBufImpl>(dev, pMem, std::move(deleter), extent)}
+        {
+        }
+
+    public:
+        std::shared_ptr<TBufImpl> m_spBufImpl;
+    };
+
+} // namespace alpaka
